% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{arima2.plot}
\alias{arima2.plot}
\title{Plot \code{Arima2} object}
\usage{
arima2.plot(x, title = NULL, tick.lab = NULL)
}
\arguments{
\item{x}{An \code{Arima2} object.}

\item{title}{Title of plot}

\item{tick.lab}{Time vector}
}
\value{
plot
}
\description{
Plot \code{Arima2} object
}
\examples{
mod <- arima2.plot(arima2(lh, order = c(1,0,1)))
}
