% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima2Table.R
\name{aicTable}
\alias{aicTable}
\title{Construct table of AIC for all combinations 0<=p<=P and 0<=q<=Q}
\usage{
aicTable(data, P, Q, D = 0)
}
\arguments{
\item{data}{a time series object}

\item{P}{a positive integer value}

\item{Q}{a positive integer value}
}
\description{
Construct table of AIC for all combinations 0<=p<=P and 0<=q<=Q
}
