% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.arima2}
\alias{plot.arima2}
\title{Plot \code{Arima2} object}
\usage{
\method{plot}{arima2}(x, roots = FALSE, title = NULL, tick.lab = NULL, ...)
}
\arguments{
\item{x}{An \code{Arima2} object. This parameter could be an object created using the function \code{arima2()}.}

\item{roots}{Would you instead prefer to plot the roots on a unit circle? Insert logical type here.}

\item{title}{Title of plot}

\item{tick.lab}{Time vector of numeric or character/string type.}

\item{...}{Other parameters}
}
\value{
plot
}
\description{
Plot \code{Arima2} object
}
\examples{
mod <- plot.arima2(arima2(lh, order = c(1,0,1)))
}
