% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.arima2}
\alias{plot.arima2}
\title{Plot \code{Arima2} object}
\usage{
\method{plot}{arima2}(x, title = NULL, tick.lab = NULL, ...)
}
\arguments{
\item{x}{An \code{Arima2} object.}

\item{title}{Title of plot}

\item{tick.lab}{Time vector}

\item{...}{Other parameters}
}
\value{
plot
}
\description{
Plot \code{Arima2} object
}
\examples{
mod <- plot.arima2(arima2(lh, order = c(1,0,1)))
}
