% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Arima2}
\alias{plot.Arima2}
\title{Plot \code{Arima2} object}
\usage{
\method{plot}{Arima2}(object)
}
\arguments{
\item{object}{An \code{Arima2} object.}
}
\description{
This function plots an \code{Arima2} object.
}
\details{
TODO: This could plot data, data + sims, or roots, roots + data, roots + data + sims.
}
\examples{
mod <- plot(arima2(lh, order = c(1,0,1)))
}
